AST: Fix mangling for nested generic types
When mangling a nominal type with a parent, we would proceed
to mangle the parent type as a context.

However, we want to disambiguate the mangling for when the
same parent type has different generic parameters bound,
for example 'A<Int>.B' vs 'A<String>.B'.

Now, if the parent type is specialized, we mangle the context
as a bound generic type.

A simple SILGen test for nested generics exercises the
mangling, and ensures that Sema and SILGen do not crash with
nested generics.

More detailed SILGen tests, as well as IRGen support for
nested generics is coming soon.