Sema: Improve checking of associated type superclass constraints
For some reason, we had a separate code path where the generic
signature of the protocol would be validated with the Self
substitution of a conforming type.

This no longer works when we minimize generic signatures, because
all associated type requirements of the protocol are eliminated.

However associated type inference could already do most of the
checks, all that was missing was checking the superclass, so
let's do it that way and remove a silly bit of code.