Parser: Better error recovery for invalid generic parameter lists
If a generic parameter list failed to parse completely, still return
it unless it has no generic parameters at all. Otherwise, we end up
with orphan GenericTypeParamDecls floating around in the parser's
scope map.

The old scope map is going away, but for proper operation with
new-style ASTScopes, we still want to set the RAngleLoc correctly.