Sema: Fix typealias lookup from protocol extension corner case
A protocol extension can add additional generic constraints on
'Self' or associated types thereof. In particular, 'Self' itself
can have a superclass constraint placed on it.

This makes members of the superclass visible from inside the
protocol extension.

Unqualified lookup into a protocol extension works by walking
the generic signature and visiting all types that appear on
the right hand side of a 'Self' constraint.

There was a bug in this logic, and classes concretely conformed
to by the superclass were not visited.