Sema: Fix UnresolvedSpecializeExpr for generic typealiases
If we have a generic typealias that was not folded down to a
TypeExpr in preCheckExpression(), either because the parent
type is an unbound generic type, or it was found by overload
resolution (ugh...), we would fail to correctly apply generic
arguments.

The problem was that CSGen would take the subexpression
of the UnresolvedSpecializeExpr and cast it to a BoundGenericType,
and apply the arguments there. But this is wrong because the
type of the subexpression type is the opened type of the
declaration. If the declaration is a type alias, the opened
type is the underlying type of the alias.

Instead, when opening an unbound generic type, record the
mapping from generic parameters to type variables in terms of
the original type declaration's generic signature.

This allows the generic arguments of a type alias to be
correctly applied.

This has been broken since Swift 3.0, and it's the last major
problem I'm aware of with generic typealiases...