Sema: Fix crash when resolving bound generic type with unbound parent
Eg, if both Foo and Bar are generic, Foo.Bar<Int> would crash.

This was introduced when I refactored the code a bit to combine
the code path for nominal types and generic typealiases. I was
a little too eager in merging them, because in fact the
substitution-based approach doesn't work when the parent type
is an unbound generic, which can occur in the nominal type case.

Fixes <https://bugs.swift.org/browse/SR-5600>,
<rdar://problem/33655028>.