Sema: Stop using FunctionType::getOld() when generating constraints for UnresolvedDotExpr
Previously we would generate the following constraint here, where
'base' was a concrete type and $arg and $result are type variables:

- ValueMember(base, $arg -> $result)

The problem is that we built a function type where the entire input
was a type variable, which would then bind to an argument list.

We could instead generate this constraint:

- ValueMember(base, $method)

That is not NFC though because we no longer have a type variable
bound to the argument list. Since $arg was created with
TVO_PreferSubtypeBinding, we would rank solutions by comparing
argument lists.

So we use the new FunctionInput constraint to extract the input
type of $method:

- ValueMember(base, $method)
- FunctionInput($method, $arg)

Once the ValueMember constraint is solved and $method has a fixed
type, $arg is bound to the function type's argument list, and
since it was created with TVO_PreferSubtypeBinding it participates
in ranking as before.

Amusingly enough, none of the tests in our suite exercised this
behavior, but the standard library would fail to build. So add a test.