Sema: Remove special handling of single 'Any' parameter
This was an artifact of some other Swift 3 compatibility behavior that has
since been removed. However, just removing it changes ranking behavior,
because we were also (unintentionally) bypassing the 'convert to Any'
penalty in this code path.

Preserve the old behavior by looking at the locator. Yes, this is really
bad and we should rethink the ranking rules here, but I'm trying to avoid
changing too many things at once with these refactorings.