Sema: Remove obsolete noescape parameter use diagnostics
The constraint solver rejects attempts to bind noescape parameters
to type variables, so this code was dead.

Note that this now allows you to assign a value of noescape function
type to a local variable with an inferred type. However that is fine,
since escaping uses of the local variable's value will themselves be
diagnosed.