AST: Avoid iterating over FileUnits in a ModuleDecl
These operations are called frequently, and would get rather expensive when
a ModuleDecl contains a large number of SourceFiles:

- lookupValue()
- lookupVisibleDecls()
- lookupClassMember()
- lookupClassMembers()
- getImportedModules()
- getImportedModulesForLookup()

For the first four, add fast paths to handle the case where all file units in
the module are SourceFiles, by refactoring SourceFile::LookupCache from a
per-file to a per-module cache.

For the last two, just always cache the result, and filter the resulting list
to ensure it only contains unique imports. With large modules, this
significantly reduces the size of the import list, since it a common
occurrence is that almost every source file in a module imports Foundation
or UIKit or whatever.