RequirementMachine: Fix subtle bug in the completion procedure
If we delete a rule because it's left hand side can be reduced using
a newly-added rule, we still have to check for overlap between the
deleted rule and all other rules.

Remove the "is deleted" checks when removing a rule pair from the
worklist, and instead, perform the check when adding the potential
overlaps from a new rule to the worklist.

Also, add some print debugging statements to the completion procedure,
enabled when RewriteSystem::DebugCompletion is set to true (which for
now at least, requires a recompilation).