RequirementMachine: Compute critical pairs directly
Previously if the left hand side of two rules overlapped, we would
compute the overlapped term and apply both rules to obtain a
critical pair.

But it is actually possible to compute the critical pair directly.
For now this has no effect other than possibly being more efficient,
but for concrete type terms we will need this formulation for the
completion procedure to work.