AssociatedTypeInference: Make a cycle breaking hack more principled
We would return an ErrorType without diagnosing anything from
NormalProtocolConformance::getTypeWitnessAndDecl() to signal
a transient condition to associated type inference. But other
callers that receive an ErrorType don't expect that this can
happen, and they in turn won't diagnose anything. So we'd end
up in SILGen with a broken AST.

Instead, push the cycle break up into getWitnessTypeForMatching(),
and remove the hack from getTypeWitnessAndDecl(). If someone
else calls getTypeWitnessAndDecl() and that causes a request
cycle, we will now diagnose the request cycle via the usual
mechanism.

Fixes rdar://120388028.